+++ packages/components/organisms/f-footer/test-utils/component-objects/f-footer.selectors.js
@@ -1,0 +1,3 @@
+export const FOOTER_COMPONENT = '[data-test-id="footer-component"]'
+export const FOOTER_ICONS = '[data-test-id="footerIcon"]'
+export const COURIER_LINKS = '[data-test-id="courierLinks-wrapper"]'