+++ packages/components/organisms/f-header/src/components/Header.vue
@@ -99,7 +99,7 @@ export default {
 
         showOffersLink: {
             type: Boolean,
-            default: true
+            default: false
         },
 
         showSkipLink: {